Move NewsTab to separate directory
