+++ src/main/java/com/atlauncher/gui/tabs/news/NewsTab.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs;
+package com.atlauncher.gui.tabs.news;
 
 import java.awt.BorderLayout;
 import java.awt.Toolkit;
@@ -33,6 +33,7 @@ import javax.swing.event.HyperlinkEvent;
 import javax.swing.text.html.HTMLEditorKit;
 import javax.swing.text.html.StyleSheet;
 
+import com.atlauncher.gui.tabs.Tab;
 import org.mini2Dx.gettext.GetText;
 
 import com.atlauncher.managers.NewsManager;
